feat: reorganize the utils
