+++ flask_fullstack/siox/__init__.py
@@ -4,4 +4,3 @@ from .eventor import EventController, EventGroupBase, EventGroupBaseMixedIn
 from .eventor import EventGroup  # DEPRECATED
 from .eventor import Namespace, SocketIO
 from .structures import EventException
-from .utils import remove_none, render_model, unpack_params, render_packed, kebabify_model
